Task 3.1 smart_home add measurement_list

+++ 3.1-drf-intro/smart_home/measurement/models.py
@@ -8,6 +8,6 @@ class Sensor(models.Model):
 
 class Measurement(models.Model):
     sensor_id = models.ForeignKey(Sensor, on_delete = models.CASCADE, related_name='mesurements')
-    temperature = models.IntegerField()
+    temperature = models.DecimalField(max_digits = 3, decimal_places=1)
     created_at = models.DateTimeField(auto_now_add = True) #автозаполнение даты и времени создания записи
